requirements.txt
# 核心库
numpy>=1.20.0
pandas>=1.3.0
requests>=2.25.0
pydantic>=2.0.0
httpx>=0.24.0

# 向量数据库和嵌入
faiss-cpu>=1.7.4; platform_system != "Darwin"
# macOS用户请使用conda安装faiss: conda install -c conda-forge faiss-cpu
transformers>=4.30.0

# LLM和大模型
dashscope>=1.10.0
openai>=1.0.0
langchain>=0.0.267
aliyunsdkcore>=2.13.36

# RSS解析
feedparser>=6.0.0
beautifulsoup4>=4.9.0

# Web应用
Flask>=2.2.0
Flask-Cors>=3.0.10
waitress>=2.1.2
markupsafe>=2.0.0

# 工具和帮助
tqdm>=4.62.0
python-dotenv>=0.19.0 